change fs_file interface to return number of bytes read/written.  this allows us to optimize out some calls to segment_get_nbytes
git-svn-id: svn+ssh://market.scs.stanford.edu/home/u1/asbestos/repo/jos64/trunk@1415 927ebb28-5305-0410-b3ca-c4b640df69b7

+++ inc/fs.h
@@ -59,8 +59,8 @@ int  fs_create(struct fs_inode dir, const char *fn, struct fs_inode *f);
 int  fs_remove(struct fs_inode dir, const char *fn, struct fs_inode f);
 int  fs_link(struct fs_inode dir, const char *fn, struct fs_inode f);
 
-int  fs_pwrite(struct fs_inode f, const void *buf, uint64_t count, uint64_t off);
-int  fs_pread(struct fs_inode f, void *buf, uint64_t count, uint64_t off);
+ssize_t fs_pwrite(struct fs_inode f, const void *buf, uint64_t count, uint64_t off);
+ssize_t fs_pread(struct fs_inode f, void *buf, uint64_t count, uint64_t off);
 int  fs_resize(struct fs_inode f, uint64_t len);
 int  fs_getsize(struct fs_inode f, uint64_t *len);
